requirements.txt
# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# PDF Processing
pdf2image>=1.16.0
Pillow>=9.0.0
img2pdf>=0.4.0

# Environment and utilities
python-dotenv>=1.0.0